Remove 2D diags from zonal_mean_2d - Move `utils._add_colormap()` to `lat_lon_plot.py`
